refactoring to support Geochron export of projects
