+++ src/main/java/org/earthtime/archivingTools/GeochronAliquotManager.java
@@ -16,6 +16,7 @@
 package org.earthtime.archivingTools;
 
 import java.awt.Color;
+import java.awt.Desktop;
 import java.awt.Graphics;
 import java.awt.Graphics2D;
 import java.awt.RenderingHints;
@@ -23,22 +24,32 @@ import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 import java.awt.event.KeyAdapter;
 import java.awt.event.KeyEvent;
+import java.io.File;
+import java.io.IOException;
+import java.util.Map;
 import java.util.Vector;
 import javax.swing.InputVerifier;
 import javax.swing.JButton;
+import javax.swing.JCheckBox;
 import javax.swing.JComponent;
 import javax.swing.JLabel;
+import javax.swing.JLayeredPane;
 import javax.swing.JOptionPane;
 import javax.swing.JPanel;
 import javax.swing.JTextField;
 import org.earthtime.UPb_Redux.ReduxConstants;
 import org.earthtime.UPb_Redux.aliquots.Aliquot;
+import org.earthtime.UPb_Redux.dateInterpretation.DateProbabilityDensityPanel;
+import org.earthtime.UPb_Redux.dateInterpretation.concordia.ConcordiaGraphPanel;
+import org.earthtime.UPb_Redux.dateInterpretation.graphPersistence.GraphAxesSetup;
 import org.earthtime.UPb_Redux.dialogs.DialogEditor;
+import org.earthtime.UPb_Redux.reduxLabData.ReduxLabData;
 import org.earthtime.UPb_Redux.samples.Sample;
 import org.earthtime.UPb_Redux.samples.SampleI;
 import org.earthtime.archivingTools.forSESAR.SesarSample;
 import org.earthtime.archivingTools.forSESAR.SesarSampleManager;
 import org.earthtime.beans.ET_JButton;
+import org.earthtime.dataDictionaries.RadDates;
 import org.earthtime.projects.ProjectI;
 
 /**
@@ -77,6 +88,12 @@ public class GeochronAliquotManager extends JPanel {
     private String[] aliquotIGSNs;
     private JButton[] registerNewAliquotButtons;
     private JButton[] viewAliquotRecordButtons;
+    private JLabel[] childStatusLabels;
+    private JButton[] showConcordiaButtons;
+    private JButton[] showPDFButtons;
+    private JCheckBox []publicOptionCheckBoxes;
+    
+    
 
     public GeochronAliquotManager(ProjectI project, SampleI sample, String userName, String password, String userCode, int x, int y, int width, int height) {
         this.project = project;
@@ -160,7 +177,7 @@ public class GeochronAliquotManager extends JPanel {
         xMarkForInValidSampleIGSN.setBounds(cumulativeWidth, TOP_MARGIN, 35, 25);
         xMarkForInValidSampleIGSN.setVisible(false);
         add(xMarkForInValidSampleIGSN);
-        cumulativeWidth += 35;
+        cumulativeWidth += 40;
 
         // next two occupy same space and show depending on condition
         registerNewSampleButton = new ET_JButton("Register New Sample");
@@ -238,6 +255,8 @@ public class GeochronAliquotManager extends JPanel {
         registerNewAliquotButtons = new JButton[aliquotCount];
         viewAliquotRecordButtons = new JButton[aliquotCount];
         sesarAliquots = new SesarSample[aliquotCount];
+        childStatusLabels = new JLabel[aliquotCount];
+        showConcordiaButtons = new JButton[aliquotCount];
 
         for (int i = 0; i < aliquotCount; i++) {
             System.out.println(sample.getSampleName() + "  >  " + activeAliquots.get(i).getAliquotName());
@@ -307,7 +326,7 @@ public class GeochronAliquotManager extends JPanel {
             xMarkForInValidAliqutIGSNs[i].setBounds(cumulativeWidth, TOP_MARGIN + 30 * (i + 1), 35, 25);
             xMarkForInValidAliqutIGSNs[i].setVisible(false);
             add(xMarkForInValidAliqutIGSNs[i]);
-            cumulativeWidth += 35;
+            cumulativeWidth += 40;
 
             // next two occupy same space and show depending on condition
             registerNewAliquotButtons[i] = new ET_JButton("Register New Aliquot");
@@ -351,12 +370,46 @@ public class GeochronAliquotManager extends JPanel {
                             JOptionPane.WARNING_MESSAGE);
                 }
             });
-            cumulativeWidth += 135;
+            cumulativeWidth += 140;
 
+            childStatusLabels[i] = new JLabel("Aliquot IGSN is not a valid child of Sample IGSN.");
+            childStatusLabels[i].setBounds(cumulativeWidth, TOP_MARGIN + 30 * (i + 1), 300, 25);
+            childStatusLabels[i].setFont(ReduxConstants.sansSerif_12_Bold);
+            childStatusLabels[i].setForeground(Color.red);
+            add(childStatusLabels[i]);
+            //cumulativeWidth += 202;
+           
+            showConcordiaButtons[i] = new ET_JButton("View Concordia in browser");
+            showConcordiaButtons[i].setBounds(cumulativeWidth, TOP_MARGIN + 30 * (i + 1), 150, 25);
+            showConcordiaButtons[i].setFont(ReduxConstants.sansSerif_10_Bold);
+            showConcordiaButtons[i].setVisible(false);
+            add(showConcordiaButtons[i], JLayeredPane.DEFAULT_LAYER);
+            showConcordiaButtons[i].addActionListener((ActionEvent e) -> {
+                produceConcordiaGraph(sample);
+            });
+
+//            JButton showPDFButton = new ET_JButton("View PDF in browser");
+//            showPDFButton.setBounds(leftMargin + 150 + 100 + 50 + 2 + 150 + 2, topMarginForSampleDetails + row * 25, 150, 25);
+//            showPDFButton.setFont(ReduxConstants.sansSerif_10_Bold);
+//            showPDFButton.setName(String.valueOf(row));
+//            showPDFButton.setVisible(false);
+//            sampleShowPDFButtons.add(showPDFButton);
+//            exportManagerLayeredPane.add(showPDFButton, JLayeredPane.DEFAULT_LAYER);
+//            showPDFButton.addActionListener((ActionEvent e) -> {
+//                EarthTimeSerializedFileInterface deserializedFile = //
+//                        (EarthTimeSerializedFileInterface) ETSerializer.GetSerializedObjectFromFile(ss.getSampleReduxFilePath().toString());
+//                Sample sample = (Sample) deserializedFile;
+//                producePDFImage(sample);
+//            });
 //
-//            
-//            
-//            
+//            JCheckBox publicOptionCheckBox = new JCheckBox("Public ?");
+//            publicOptionCheckBox.setBounds(leftMargin + 150 + 100 + 50 + 2 + 150 + 2 + 150 + 2, topMarginForSampleDetails + row * 25, 100, 25);
+//            publicOptionCheckBox.setFont(ReduxConstants.sansSerif_10_Bold);
+//            publicOptionCheckBox.setName(String.valueOf(row));
+//            publicOptionCheckBox.setVisible(false);
+//            samplePublicCheckBoxes.add(publicOptionCheckBox);
+//            exportManagerLayeredPane.add(publicOptionCheckBox, JLayeredPane.DEFAULT_LAYER);
+////           
 //            
 //            
 //            
@@ -435,6 +488,10 @@ public class GeochronAliquotManager extends JPanel {
 
         viewAliquotRecordButtons[index].setVisible(valid);
         registerNewAliquotButtons[index].setVisible(!valid);
+
+        childStatusLabels[index].setVisible(!valid);
+        showConcordiaButtons[index].setVisible(valid);
+
 //
 //        saveButton.setVisible(valid);
         //uploadButton.setVisible(valid);
@@ -496,6 +553,118 @@ public class GeochronAliquotManager extends JPanel {
         }
     }
 
+        private void produceConcordiaGraph(SampleI sample) {
+        // feb 2015 code copied and modified from aliquot manager for user interface prototyping
+        // TODO: refactor both locations to smaple and make more robust
+        // TODO: use create virtual file system
+
+//        try {
+//            FileSystem virtualFileSystem = Jimfs.newFileSystem(Configuration.unix());
+//            Path virtualPathToSVG = virtualFileSystem.getPath("", sample.getSampleName() + ".svg");
+//
+//        } catch (Exception e) {
+//            System.out.println("JIMFS Worked");
+//        }
+        File tempConcordiaSVG = new File(sample.getSampleName() + "_tempConcordia.svg");
+        sample.setMyReduxLabData(ReduxLabData.getInstance());
+
+        ConcordiaGraphPanel concordiaGraphPanel = new ConcordiaGraphPanel((Sample)sample, null);
+
+        sample.getSampleDateInterpretationGUISettings().//
+                setConcordiaOptions(concordiaGraphPanel.getConcordiaOptions());
+        concordiaGraphPanel.//
+                setFadedDeselectedFractions(false);
+
+        // set choices per options code copied (TODO: REFACTOR ME) from SampleDateInterpretations
+        Map<String, String> CGO = concordiaGraphPanel.getConcordiaOptions();
+        if (CGO.containsKey("showEllipseLabels")) {
+            concordiaGraphPanel.setShowEllipseLabels(false);
+        }
+        if (CGO.containsKey("showExcludedEllipses")) {
+            concordiaGraphPanel.setShowExcludedEllipses(true);
+        }
+
+        concordiaGraphPanel.setSelectedFractions(sample.getUPbFractions());
+
+        concordiaGraphPanel.setBounds(510, 0, 580, 405);
+        concordiaGraphPanel.setCurrentGraphAxesSetup(new GraphAxesSetup("C", 2));
+        concordiaGraphPanel.setGraphWidth(565 - GraphAxesSetup.DEFAULT_GRAPH_LEFT_MARGIN_VERTICAL_LABELS);
+        concordiaGraphPanel.setGraphHeight(385);
+
+        concordiaGraphPanel.setYorkFitLine(null);
+        concordiaGraphPanel.getDeSelectedFractions().clear();
+        concordiaGraphPanel.setPreferredDatePanel(null);
+
+        concordiaGraphPanel.setShowTightToEdges(true);
+
+        concordiaGraphPanel.refreshPanel();
+
+        concordiaGraphPanel.setShowTightToEdges(false);
+
+        boolean saveShowTitleBox = concordiaGraphPanel.isShowTitleBox();
+        concordiaGraphPanel.setShowTitleBox(false);
+        concordiaGraphPanel.setUploadToGeochronMode(true);
+
+        concordiaGraphPanel.outputToSVG(tempConcordiaSVG);
+        // concordiaGraphPanel.outputToSVG(virtualPathToSVG.toFile());
+
+        concordiaGraphPanel.setShowTitleBox(saveShowTitleBox);
+        concordiaGraphPanel.setUploadToGeochronMode(false);
+
+        // show in a browser
+//            Desktop.getDesktop().browse(virtualPathToSVG.toUri());
+        try {
+            Desktop.getDesktop().browse(tempConcordiaSVG.toURI());
+        } catch (IOException iOException) {
+            System.out.println("Browser issue " + iOException.getMessage());
+        }
+
+    }
+
+    private void producePDFImage(Sample sample) {
+        File tempProbabilitySVG = new File(sample.getSampleName() + "_tempProbabilityDensity.svg");
+        sample.setMyReduxLabData(ReduxLabData.getInstance());
+
+        DateProbabilityDensityPanel probabilityPanel = new DateProbabilityDensityPanel(sample);
+
+        // use default if user has not initialized
+        if (probabilityPanel.getSelectedFractions().isEmpty()) {
+            probabilityPanel.//
+                    setSelectedFractions(sample.getUpbFractionsUnknown());
+            probabilityPanel.//
+                    getDeSelectedFractions().clear();
+
+            probabilityPanel.setGraphWidth(565);
+            probabilityPanel.setGraphHeight(385);
+
+            probabilityPanel.setSelectedHistogramBinCount(5);
+
+            if (sample.isTypeLegacy() & sample.getAnalysisPurpose().equals(ReduxConstants.ANALYSIS_PURPOSE.DetritalSpectrum)) {
+                probabilityPanel.setChosenDateName(RadDates.bestAge.getName());
+            } else {
+                probabilityPanel.setChosenDateName(RadDates.age207_206r.getName());
+            }
+
+            probabilityPanel.showTight();
+
+        } else {
+            probabilityPanel.setGraphWidth(565);
+            probabilityPanel.setGraphHeight(385);
+        }
+
+        probabilityPanel.setUploadToGeochronMode(true);
+
+        probabilityPanel.outputToSVG(tempProbabilitySVG);
+
+        probabilityPanel.setUploadToGeochronMode(false);
+
+        try {
+            Desktop.getDesktop().browse(tempProbabilitySVG.toURI());
+        } catch (IOException iOException) {
+            System.out.println("Browser issue " + iOException.getMessage());
+        }
+
+    }
     /**
      *
      * @param g
